pyproject.toml
[project]
name = "healthcare-processing-prototype"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "nats-py>=2.7",
    "asyncpg>=0.29",
    "pydantic>=2.6",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "ollama>=0.6.1",
    "prometheus-client>=0.19",
    "opentelemetry-api>=1.0",
    "opentelemetry-sdk>=1.0",
    "opentelemetry-exporter-otlp>=0.44b0",
    "opentelemetry-instrumentation-fastapi>=0.44b0",
    "opentelemetry-instrumentation-httpx>=0.44b0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
    "mypy",
]

[dependency-groups]
debug = [
    "debugpy>=1.8",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
